BUG: revise offset correction in `BuildCalendar`
In original version, date will be corrected only when offset is not 1.
However, it should be corrected when offset is not 0 instead.
Furthermore, offset will be set to 7 when it was 0. So that there will
be a row of day in previous month set as a padding row to target month.