BUG: event handler is NULL if it is read from json
Event handler cannot be serialized into json (json should contains only
data/information). However, scheduler have to re-create handler after
the program is re-opened.
To achieve this functionality, an `EventHandlerFactory` is provided in
`Jarloo.Calendar`. So that caller can define its own workflow of event
handlers by inherit it.

Additionally, eventArgs should be saved with type. Because a custom
eventArgs cannot be casted from the basic type `EventArgs`.
To do this, methods in `Jarloo.Calendar.DataControl` is re-implemented
by `JsonSerializer`, instead of `JToken`. And set the `TypeNameHandling`
of JsonSerializer to `All`.