BUG: calendar has to be created to load events
EventManager is a part of Jarloo.Calendar. To load events when program
is started, calendar has to be created. However, it is impossible(?) to
created a user control(calendar) and load it after the
container(PageCalendar) is created.
And in this project, default page is required to be PagePlaylist. So
that we cannot just set it to PageCalendar.
A workaround for this problem is creating PagePlaylist and PageCalendar
when MainWindow is just created. So that EventManager can be loaded to
load all events. Then we switch the page to PagePlaylist. In the end, we
can keep the default page being PagePlaylist but also make EventManager
works properly.