UPDATE: in-place-editable ListBoxItem
The basic idea to implement this feature is overwrite the
controlTemplate of ListBoxItem by 2 overlapped TextBlock and TextBox.
TextBlock is used to show content, and TextBox is used to modify
content.

To implement this feature, we need to set two event:
1. "KeyDown" on ListBoxItem: using EventSetter in the style of
ListBoxItem (ItemContainerStyle of ListBox).
2. "LostFocus" on TextBox inside the controlTemplate of ListBoxItem.
(Notice that the correct sender of this event is "TextBox", instead of
"ListBoxItem". Because once the TextBox is showing and being clicked by
user, the control getting focused is "TextBox", not "ListBoxItem".)

In default, TextBlock is visible and TextBox is hidden. So that user is
not able to modify the content directly by access the TextBox. Once the
key "F2" is down, we exchange the visibility of TextBlock and TextBox.
And user can modify content of ListBoxItem in the meantime. And after
the TextBox lost focus, save the change of content.

Besides, a new property `GUID` is added into `Playlist` for searching
desired one.
Because lb_PlaylistMenu is modified to binding an
ObservableCollection<Playlist> as an ItemSource. And inside the
controlTemplate of ListBoxItem, "Text" of both TextBox and TextBlock
binds to `ListName`. So that once the text of TextBox is modified,
ObservableCollection<Playlist> will be changed too.
And it make us not able to find the selected list by its name.