MAINT: rewrite path selection of TrackDB, EventDB
`PlaylistInfoPath` is renamed to `TrackDBPath`.
In `MPLite.PageSetting`, path of `TrackDB` and `EventDB` is set by
OpenFileDialog. This approach may takes more affect to notify user that
format of files have to be JSON.
So that OpenFileDialog is replaced by FolderBrowserDialog. Now, user
only need to selected the directory to save TrackDB and EventDB.
In addition, some commented-out code is removed in this commit.