BUG: event to be shown should be retrieved from DB
1. When "Event info" is click in the context menu of calendar, event to
be shown should be retrieved from EventDB (stored in hard drive),
instead of EventDB loaded in runtime.
The original approach result in "EmptyQueueException" when the
triggering time of an event are all before `DateTime.Now`.
2. In `MultiTriggerEvent.UpdateBeginningTime()`, non-recurring event
(RecurringFrequency.None) should be considered. In original approach,
result of `base.UpdateBeginningTime()` will always be false for
non-recurring event. And it leads to failure of `base.SetTimer()`. (...
`else if (this.UpdateBeginningTime() && BeginningTime >
DateTime.Now)...`)
3. In `EventCollection.UpdateEvent()`, a statement `this.EventList =
ec.EventList;` should be added before save the EventCollection.
4. In `EventManager.DisposeActivatedTimers()`, list to be iterated
should be `ActivatedEvents` instead of `EventDB`.