BUG: trackStatus won't be cleared by event action
For a non-recurring event, status of the played track won't be cleared
after event endding.
The reason is that the track queue is cleared (by `NewSelectionEvent`)
before `StopPlayerRequestEvent`, and then in `MusicPlayer.Stop()`, it
cannot fire `PlayerStoppedEvent` with info of current playing track.
To fix this bug, `NewSelectionEvent` should be fired only after the
music player is stopped.