MAINT: clear cahce when player is stopped manually
When `StopPlayerRequestEvent` is fired, music player will be stopped.
Originally, track queue is not reset at this moment. It may lead to
search failure when user try to player a new track after the playlist is
updated. (Because cache is not cleared, player will try to reset the
status of previous track. However, once the player is updated and the
track does not exist, exception occurs.)
So now track queue will be cleared when music player is stopped by
firing `StopPlayerRequestEvent`.