BUG: conflict of key down event
After hotkeys of player controls are set, conflicts might ouccr when
user try to edit name of playlist.
To solve this problem, a lock `MPLiteSetting.IsEditting` is added.
When user is editting hotkey or name of playlist,
`MPLiteSetting.IsEditting` will be set to `true`. So that hotkey
processing in `winMain_PreviewKeyDown()` won't be triggered.

Besides, beeping sound of key press is disabled by setting `e.Handled`
to `true`. (see `winMain_PreviewKeyDown()`)