BUG: NullReferenceException
Message "Object reference not set to an instance of an object" shows up
when there is a track selected in listview.
This exception does not shows up when there are just a few tracks or
tracklists. The actual cause of this bug has not been confirmed.
Currently, a workround to avoid this exception being thrown is adding a
check statement after the type casting
(`ListView.ItemContainerGenerator.ContainerFromIndex() as
ListViewItem`).