BUG: incorrect playback order in shuffle mode
The first track was always be played first when a scheduler event, in
which the playback mode is shuffle, is triggered.

There are two causes leading to this bug:
1. There is no proper choice provided to user to randomly select first
track to play.
2. In class `TrackQueue.SetTrackQueue()`, `beginningIdx` will be
modified to 0 once there is no track selected. However, in shuffle mode,
this operation results in a fixed `beggingIdx`.

To solve this problem,
1. Track indices in comboBox should be refreshed when the selected
`PlaybackMode` is changed. If `PlaybackMode` is `Shuffle` or
`ShuffleOnce`, an additional choice `-1` will be added. Where `-1` means
"no specific track is choosed to be played first".
2. In `TrackQueue.SetTrackQueue()`, handling of `beginningIdx` is
implemented in each individual subfunctions. (e.g.
_SetTrackQueue_Default, _SetTrackQueue_RepeatList...)