BUG: EmptyPlaylistException is thrown mistakenly
`EmptyPlaylistException` would be thrown out when a scheduler event is
triggered and currently selected playlist is empty.
This exception is thrown mistakenly, because it should be thrown only
when the playlist which is going to be played is empty.
Cause of this bug is that a check statement `lv_Playlist.Item,Count ==
0` is added in `PagePlaylist.GetTrack()`.