MAINT: rewrite setting of launching at startup
In previous version, app cannot launch at startup sucessfully. Because
the value of register have to be a string with "". However, we also have
to set the working directory when app is launched to  ensure this app
can get desired info in its workspace.
To achieve this goal, we have to set current directory when the app is
just launch (see App()@App.xaml.cs).