BUG: track index of an event is not set properly
For events with playback mode of `Shuffle`, `ShuffleOnce`, the track
index to be played first will shown incorrectly in `WindowEventSetting`.
This bug results from the `cmbTrackIndex.ItemSource`. It will be changed
according to the playback mode.
In previous version, `cmbTrackIndex.SeletecIndex` is set before
`cmbPlaybackMode.SelectedItem`, and the default ItemSource of
`cmbTrackIndex` is start from 0 ~ trackAmount. However, the range of
`cmbTrackIndex` will change to -1 ~ trackAmount once the
`cmbPlaybackMode.SelectedItem` is `Shuffle` or `ShuffleOnce`.
To fix this bug, we should set playback mode before setting index of the
track to be played first.